+++ src/pythonic_fp/homepage/top_level_releases.py
@@ -7,7 +7,7 @@ These are the most resent *Pythonic FP* project PyPI releases.
 ============================ =========================== ======= =============
 PyPI and GitHub Name         Python Module               Version Date Released
 ============================ =========================== ======= =============
-pythonic-fp-booleans         pythonic_fp.booleans        2.0.0   2025-09-27
+pythonic-fp-booleans         pythonic_fp.booleans        2.1.0   2025-12-02
 pythonic-fp-circulararray    pythonic_fp.circulararray   6.0.0   2025-09-26
 pythonic-fp-containers       pythonic_fp.containers      4.0.0   2025-09-28
 pythonic-fp-fptools          pythonic_fp.fptools         5.1.2   2025-09-28